gmp: Pass include directories to Haskell compiler
Previously the include directories weren't being included in the arguments
used to invoke GHC when building the in-tree GMP. This caused the build to fail
with missing include files.

I considered for a moment using Cabal's `--extra-include-dirs` flag to pass
these directories. However, it's important that we
Instead of duplicating the include directory logic for the Haskell compiler,
I just use Cabal's --extra-include-dirs flag, ensuring that